Move snapshot related functionality to docker plugin
